# web/api-generated/models/edge-add-input.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Notebook Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EdgeData } from './edge-data';
// May contain unused imports in some cases
// @ts-ignore
import type { EdgeDataTypeEnum } from './edge-data-type-enum';
// May contain unused imports in some cases
// @ts-ignore
import type { EdgeHandleEnum } from './edge-handle-enum';

/**
 * 
 * @export
 * @interface EdgeAddInput
 */
export interface EdgeAddInput {
    /**
     * 
     * @type {number}
     * @memberof EdgeAddInput
     */
    'flowId': number;
    /**
     * 
     * @type {number}
     * @memberof EdgeAddInput
     */
    'sourceNodeId'?: number;
    /**
     * 
     * @type {number}
     * @memberof EdgeAddInput
     */
    'targetNodeId'?: number;
    /**
     * 
     * @type {EdgeHandleEnum}
     * @memberof EdgeAddInput
     */
    'sourceHandle'?: EdgeHandleEnum;
    /**
     * 
     * @type {EdgeHandleEnum}
     * @memberof EdgeAddInput
     */
    'targetHandle'?: EdgeHandleEnum;
    /**
     * 
     * @type {EdgeData}
     * @memberof EdgeAddInput
     */
    'data'?: EdgeData;
    /**
     * 
     * @type {EdgeDataTypeEnum}
     * @memberof EdgeAddInput
     */
    'dataType'?: EdgeDataTypeEnum;
}