# web/api-generated/models/edge-entity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Notebook Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EdgeData } from './edge-data';
// May contain unused imports in some cases
// @ts-ignore
import type { EdgeDataTypeEnum } from './edge-data-type-enum';
// May contain unused imports in some cases
// @ts-ignore
import type { EdgeHandleEnum } from './edge-handle-enum';

/**
 * 
 * @export
 * @interface EdgeEntity
 */
export interface EdgeEntity {
    /**
     * 
     * @type {EdgeHandleEnum}
     * @memberof EdgeEntity
     */
    'sourceHandle'?: EdgeHandleEnum;
    /**
     * 
     * @type {EdgeHandleEnum}
     * @memberof EdgeEntity
     */
    'targetHandle'?: EdgeHandleEnum;
    /**
     * 
     * @type {EdgeDataTypeEnum}
     * @memberof EdgeEntity
     */
    'dataType'?: EdgeDataTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof EdgeEntity
     */
    'id': number;
    /**
     * 
     * @type {number}
     * @memberof EdgeEntity
     */
    'flowId': number;
    /**
     * 
     * @type {number}
     * @memberof EdgeEntity
     */
    'sourceNodeId'?: number;
    /**
     * 
     * @type {number}
     * @memberof EdgeEntity
     */
    'targetNodeId'?: number;
    /**
     * 
     * @type {EdgeData}
     * @memberof EdgeEntity
     */
    'data': EdgeData;
    /**
     * 
     * @type {string}
     * @memberof EdgeEntity
     */
    'updatedAt': string;
}