# web/api-generated/models/edge-handle-enum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Notebook Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */

export const EdgeHandleEnum = {
    Top: 'Top',
    Right: 'Right',
    Bottom: 'Bottom',
    Left: 'Left'
} as const;

export type EdgeHandleEnum = typeof EdgeHandleEnum[keyof typeof EdgeHandleEnum];