# web/api-generated/models/edge-patch-input.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Notebook Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EdgeData } from './edge-data';
// May contain unused imports in some cases
// @ts-ignore
import type { EdgeDataTypeEnum } from './edge-data-type-enum';
// May contain unused imports in some cases
// @ts-ignore
import type { EdgeHandleEnum } from './edge-handle-enum';

/**
 * 
 * @export
 * @interface EdgePatchInput
 */
export interface EdgePatchInput {
    /**
     * 
     * @type {number}
     * @memberof EdgePatchInput
     */
    'sourceNodeId'?: number;
    /**
     * 
     * @type {number}
     * @memberof EdgePatchInput
     */
    'targetNodeId'?: number;
    /**
     * 
     * @type {EdgeHandleEnum}
     * @memberof EdgePatchInput
     */
    'sourceHandle'?: EdgeHandleEnum;
    /**
     * 
     * @type {EdgeHandleEnum}
     * @memberof EdgePatchInput
     */
    'targetHandle'?: EdgeHandleEnum;
    /**
     * 
     * @type {EdgeData}
     * @memberof EdgePatchInput
     */
    'data'?: EdgeData;
    /**
     * 
     * @type {EdgeDataTypeEnum}
     * @memberof EdgePatchInput
     */
    'dataType'?: EdgeDataTypeEnum;
}