# web/api-generated/models/node-entity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Notebook Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { NodeData } from './node-data';
// May contain unused imports in some cases
// @ts-ignore
import type { NodeDataTypeEnum } from './node-data-type-enum';

/**
 * 
 * @export
 * @interface NodeEntity
 */
export interface NodeEntity {
    /**
     * 
     * @type {NodeDataTypeEnum}
     * @memberof NodeEntity
     */
    'dataType'?: NodeDataTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof NodeEntity
     */
    'id': number;
    /**
     * 
     * @type {number}
     * @memberof NodeEntity
     */
    'flowId': number;
    /**
     * 
     * @type {number}
     * @memberof NodeEntity
     */
    'positionX': number;
    /**
     * 
     * @type {number}
     * @memberof NodeEntity
     */
    'positionY': number;
    /**
     * 
     * @type {NodeData}
     * @memberof NodeEntity
     */
    'data': NodeData;
    /**
     * 
     * @type {string}
     * @memberof NodeEntity
     */
    'updatedAt': string;
}